feat!: removed http endpoints
BREAKING CHANGES: removed http endpoint
