+++ internal/controller/grpc/handler.go
@@ -2,12 +2,15 @@ package grpc
 
 import (
 	"context"
+	"errors"
 	"github.com/vgekko/anistream-content/internal/controller/grpc/pb"
 	"github.com/vgekko/anistream-content/internal/entity"
 	"github.com/vgekko/anistream-content/internal/usecase"
-	"github.com/vgekko/anistream-content/pkg/logger/sl"
+	"github.com/vgekko/anistream-content/pkg/apperror"
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/reflection"
+	"google.golang.org/grpc/status"
 	"log/slog"
 )
 
@@ -26,17 +29,31 @@ func NewContentServerGrpc(gserver *grpc.Server, uc usecase.ContentUseCase, log *
 	reflection.Register(gserver)
 }
 
+type errorMap struct {
+	err  error      // domain error
+	code codes.Code // grpc status code
+}
+
 func (s *ContentServer) GetTitleContent(ctx context.Context, in *pb.GetTitleContentRequest) (*pb.GetTitleContentReply, error) {
+	possibleErrors := []errorMap{
+		{apperror.ErrTitleNotFound, codes.NotFound},
+		{apperror.ErrNoSearchParams, codes.InvalidArgument},
+		{apperror.ErrMissingOrInvalidToken, codes.Internal},
+		{apperror.ErrUnknown, codes.Unknown},
+	}
+
 	titleInfos, err := s.uc.Search(entity.TitleFilter{Opt: in.Filter.Opt, Val: in.Filter.Val})
-	if err != nil {
-		s.log.Error("grpc: GetTitleContent: ", sl.Err(err))
-		return nil, err
+	for _, pErr := range possibleErrors {
+		if errors.Is(errors.Unwrap(err), pErr.err) {
+			s.log.Error(err.Error())
+			return nil, status.Error(pErr.code, err.Error())
+		}
 	}
 
 	reply := pb.GetTitleContentReply{}
 	reply.TitleContent = transformTitleContent(titleInfos)
 
-	return &reply, nil
+	return &reply, status.New(codes.OK, "success").Err()
 }
 
 func transformTitleContent(titles []entity.TitleContent) []*pb.TitleContent {
