+++ pkg/cache/cache.go
@@ -3,14 +3,14 @@ package cache
 import (
 	"github.com/allegro/bigcache"
 	"github.com/vgekko/anistream-content/config"
-	"github.com/vgekko/anistream-content/pkg/logger/sl"
 	"log/slog"
 )
 
 func New(cfg config.Cache) *bigcache.BigCache {
 	bc, err := bigcache.NewBigCache(bigcache.DefaultConfig(cfg.TTL))
 	if err != nil {
-		slog.Error("cannot initialize cache: ", sl.Err(err))
+		slog.Error("failed to initialize bigcache")
+		return nil
 	}
 
 	return bc
