--- pkg/httpserver/server.go
@@ -1,57 +1,0 @@
-package httpserver
-
-import (
-	"context"
-	"net/http"
-	"time"
-)
-
-const (
-	defaultReadTimeout     = 5 * time.Second
-	defaultWriteTimeout    = 5 * time.Second
-	defaultAddr            = ":8800"
-	defaultShutdownTimeout = 3 * time.Second
-)
-
-type Server struct {
-	server          *http.Server
-	notify          chan error
-	shutdownTimeout time.Duration
-}
-
-func Start(handler http.Handler) *Server {
-	httpServer := &http.Server{
-		Addr:         defaultAddr,
-		ReadTimeout:  defaultReadTimeout,
-		WriteTimeout: defaultWriteTimeout,
-		Handler:      handler,
-	}
-
-	s := &Server{
-		server:          httpServer,
-		notify:          make(chan error, 1),
-		shutdownTimeout: defaultShutdownTimeout,
-	}
-
-	s.start()
-
-	return s
-}
-
-func (s *Server) start() {
-	go func() {
-		s.notify <- s.server.ListenAndServe()
-		close(s.notify)
-	}()
-}
-
-func (s *Server) Notify() <-chan error {
-	return s.notify
-}
-
-func (s *Server) Shutdown() error {
-	ctx, cancel := context.WithTimeout(context.Background(), s.shutdownTimeout)
-	defer cancel()
-
-	return s.server.Shutdown(ctx)
-}
