+++ proto/content.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 option go_package = "./internal/controller/grpc/pb";
 
 service ContentService {
-  rpc GetTitleContent (GetTitleContentRequest) returns (GetTitleContentReply);
+  rpc GetTitleContent (GetTitleContentRequest) returns (GetTitleContentReply); //{
 }
 
 message TitleContent {
